requirement.txt
flask-cors
Flask
groq
python-docx
python-dotenv

remark-gfm
react
react-markdown
